# docs/swagger.yaml
basePath: /
definitions:
  ResponseDatasource:
    properties:
      data:
        items:
          type: integer
        type: array
      name:
        type: string
    type: object
  properErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Using a translation service as an example
  title: interview-crud-files api
  version: "1.0"
paths:
  /list:
    get:
      description: Read datasource list
      operationId: readDatasourceList
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Incorrect request
          schema:
            $ref: '#/definitions/properErrorResponse'
        "403":
          description: Permission deny
          schema:
            $ref: '#/definitions/properErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/properErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/properErrorResponse'
      summary: Read datasource list
      tags:
      - datasource list
  /source:
    delete:
      description: delete datasource by name
      operationId: datasource_delete
      parameters:
      - description: delete by name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "403":
          description: No data or no access to it
          schema:
            $ref: '#/definitions/properErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/properErrorResponse'
      summary: delete datasource
      tags:
      - datasource
    get:
      description: read datasource
      operationId: readDatasource
      parameters:
      - description: name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ResponseDatasource'
        "400":
          description: Incorrect request
          schema:
            $ref: '#/definitions/properErrorResponse'
        "403":
          description: Permission deny
          schema:
            $ref: '#/definitions/properErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/properErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/properErrorResponse'
      summary: read datasource
      tags:
      - datasource
    patch:
      consumes:
      - multipart/form-data
      description: update datasource by name
      operationId: datasource_update
      parameters:
      - description: files to update
        in: formData
        items:
          type: file
        name: files
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Incorrect request
          schema:
            $ref: '#/definitions/properErrorResponse'
        "403":
          description: Permission deny
          schema:
            $ref: '#/definitions/properErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/properErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/properErrorResponse'
      summary: update datasource
      tags:
      - datasource
    post:
      consumes:
      - multipart/form-data
      description: create datasource by datasource name
      operationId: createDatasource
      parameters:
      - description: files to download
        in: formData
        items:
          type: file
        name: files
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Incorrect request
          schema:
            $ref: '#/definitions/properErrorResponse'
        "403":
          description: Permission deny
          schema:
            $ref: '#/definitions/properErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/properErrorResponse'
      summary: create datasource
      tags:
      - datasource
swagger: "2.0"